New clean and shiny architecture
This commit makes our architecture more simple and clean.
It introduces:
- PhantomPage - host page for the main execution context
- CustomPage moved to its own file (header + source)
- Config renamed to Settings
- QtWebKit dropped support for old compilers, so we can safely use precompiled headers - config.h
- decoupled page renderer. I moved page renderer to its own class to solve problems
with printing a web page to PDF. With this new class we can use clipRect for PDF

This architecture is more stable and consumes less memory because
of correct QObject relations. All nested (dependent) object
will be deleted correctly.
