+++ src/env.h
@@ -1,4 +1,4 @@
-/*
+﻿/*
   This file is part of the PhantomJS project from Ofi Labs.
 
   Copyright (C) 2012 execjosh, http://execjosh.blogspot.com
@@ -27,8 +27,7 @@
   THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-#ifndef ENV_H
-#define ENV_H
+#pragma once
 
 #include <QObject>
 #include <QVariantMap>
@@ -47,5 +46,3 @@ private:
 
     QVariantMap m_map;
 };
-
-#endif // ENV_H
