+++ src/modules/webpage.js
@@ -1,4 +1,4 @@
-/*
+﻿/*
   This file is part of the PhantomJS project from Ofi Labs.
 
   Copyright (C) 2011 Ariya Hidayat <ariya.hidayat@gmail.com>
@@ -158,7 +158,8 @@ function definePageCallbackHandler(page, handlers, handlerName, callbackConstruc
       return (!!handlerObj && typeof handlerObj.callback === 'function' && typeof handlerObj.connector === 'function') ?
         handlers[handlerName].callback :
         undefined;
-    }
+    },
+    configurable: true
   });
 }
