+++ src/phantomjs.qrc
@@ -2,7 +2,6 @@
     <qresource prefix="/">
         <file>phantomjs-icon.png</file>
         <file>bootstrap.js</file>
-        <file>configurator.js</file>
         <file>modules/webpage.js</file>
         <file>modules/webserver.js</file>
         <file>modules/fs.js</file>
