--- src/phantomjs_os2.rc
@@ -1,1 +1,0 @@
-ICON 1 "phantomjs_os2.ico"
