+++ src/phantompage.h
@@ -1,0 +1,74 @@
+﻿/*
+  This file is part of the PhantomJS project from Ofi Labs.
+
+  Copyright (C) 2011 Ariya Hidayat <ariya.hidayat@gmail.com>
+  Copyright (C) 2011 Ivan De Marino <ivan.de.marino@gmail.com>
+  Copyright (C) 2017 Vitaly Slobodin <vitaliy.slobodin@gmail.com>
+
+  Redistribution and use in source and binary forms, with or without
+  modification, are permitted provided that the following conditions are met:
+
+    * Redistributions of source code must retain the above copyright
+      notice, this list of conditions and the following disclaimer.
+    * Redistributions in binary form must reproduce the above copyright
+      notice, this list of conditions and the following disclaimer in the
+      documentation and/or other materials provided with the distribution.
+    * Neither the name of the <organization> nor the
+      names of its contributors may be used to endorse or promote products
+      derived from this software without specific prior written permission.
+
+  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+  ARE DISCLAIMED. IN NO EVENT SHALL <COPYRIGHT HOLDER> BE LIABLE FOR ANY
+  DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+  (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+  LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+  ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
+  THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+
+#pragma once
+
+#include <QPointer>
+#include <QWebInspector>
+#include <QWebPage>
+
+class PhantomPage : public QObject {
+    Q_OBJECT
+
+public:
+    explicit PhantomPage(QObject* parent = Q_NULLPTR);
+    ~PhantomPage();
+
+    void bootstrap();
+
+    QVariant execute(const QString& source);
+    QVariant executeFromFile(const QString& file);
+
+    QWebFrame* mainFrame() const;
+
+    void showInspector(qint16 port);
+
+Q_SIGNALS:
+    void consoleMessageReceived(const QString& message,
+                          int lineNumber,
+                          const QString& sourceID);
+    void javaScriptErrorSent(QString message);
+    void initialized(bool ok);
+
+private Q_SLOTS:
+    void onConsoleMessageReceived(QWebPage::MessageSource source,
+                                  QWebPage::MessageLevel level,
+                                  const QString& message,
+                                  int lineNumber,
+                                  const QString& sourceID);
+    void onLoadFinished(bool status);
+
+private:
+    QPointer<QWebPage> m_webPage;
+    QPointer<QWebInspector> m_webInspector;
+
+    Q_DISABLE_COPY(PhantomPage)
+};
