+++ src/qt-qpa-platform-plugin/phantom.pri
@@ -1,4 +1,6 @@
-QT += core-private gui-private platformsupport-private
+QT += \
+    core-private gui-private \
+    eventdispatcher_support-private fontdatabase_support-private
 
 SOURCES += $$PWD/phantomintegration.cpp \
            $$PWD/phantombackingstore.cpp
