+++ src/webpage.cpp
@@ -28,7 +28,6 @@
   THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-#include "webpage.h"
 
 #include <QBuffer>
 #include <QDebug>
@@ -53,16 +52,12 @@
 #include "callback.h"
 #include "config.h"
 #include "consts.h"
-#include "cookiejar.h"
-#include "networkaccessmanager.h"
+#include "custompage.h"
 #include "phantom.h"
 #include "system.h"
-#include "utils.h"
-
-#ifdef Q_OS_WIN
-#include <fcntl.h>
-#include <io.h>
-#endif
+#include "webpage.h"
+#include "network/cookiejar.h"
+#include "network/networkaccessmanager.h"
 
 // Ensure we have at least head and body.
 #define BLANK_HTML                      "<html><head></head><body></body></html>"
@@ -70,193 +65,6 @@
 #define INPAGE_CALL_NAME                "window.callPhantom"
 #define CALLBACKS_OBJECT_INJECTION      INPAGE_CALL_NAME" = function() { return window." CALLBACKS_OBJECT_NAME ".call.call(_phantom, Array.prototype.slice.call(arguments, 0)); };"
 #define CALLBACKS_OBJECT_PRESENT        "typeof(window." CALLBACKS_OBJECT_NAME ") !== \"undefined\";"
-
-#define STDOUT_FILENAME "/dev/stdout"
-#define STDERR_FILENAME "/dev/stderr"
-
-
-/**
-  * @class CustomPage
-  */
-class CustomPage: public QWebPage
-{
-    Q_OBJECT
-
-public:
-    CustomPage(WebPage* parent = 0)
-        : QWebPage(parent)
-        , m_webPage(parent)
-    {
-        m_userAgent = QWebPage::userAgentForUrl(QUrl());
-        setForwardUnsupportedContent(true);
-    }
-
-    bool extension(Extension extension, const ExtensionOption* option, ExtensionReturn* output)
-    {
-        Q_UNUSED(option);
-
-        if (extension == ChooseMultipleFilesExtension) {
-            static_cast<ChooseMultipleFilesExtensionReturn*>(output)->fileNames = m_uploadFiles;
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    void setCookieJar(CookieJar* cookieJar)
-    {
-        m_cookieJar = cookieJar;
-    }
-
-public slots:
-    bool shouldInterruptJavaScript()
-    {
-        m_webPage->javascriptInterrupt();
-
-        if (m_webPage->m_shouldInterruptJs) {
-
-            // reset our flag
-            m_webPage->m_shouldInterruptJs = false;
-            return true;
-        }
-        return false;
-    }
-
-protected:
-    bool supportsExtension(Extension extension) const
-    {
-        return extension == ChooseMultipleFilesExtension;
-    }
-
-    QString chooseFile(QWebFrame* originatingFrame, const QString& oldFile)
-    {
-        Q_UNUSED(originatingFrame);
-
-        // Check if User set a file via File Picker
-        QString chosenFile = m_webPage->filePicker(oldFile);
-        if (chosenFile == QString::null && m_uploadFiles.count() > 0) {
-            // Check if instead User set a file via uploadFile API
-            chosenFile = m_uploadFiles.first();
-        }
-
-        // Return the value coming from the "filePicker" callback, IFF not null.
-        qDebug() << "CustomPage - file chosen for upload:" << chosenFile;
-        return chosenFile;
-    }
-
-    void javaScriptAlert(QWebFrame* originatingFrame, const QString& msg)
-    {
-        Q_UNUSED(originatingFrame);
-        emit m_webPage->javaScriptAlertSent(msg);
-    }
-
-    bool javaScriptConfirm(QWebFrame* originatingFrame, const QString& msg)
-    {
-        Q_UNUSED(originatingFrame);
-        return m_webPage->javaScriptConfirm(msg);
-    }
-
-    bool javaScriptPrompt(QWebFrame* originatingFrame, const QString& msg, const QString& defaultValue, QString* result)
-    {
-        Q_UNUSED(originatingFrame);
-        return m_webPage->javaScriptPrompt(msg, defaultValue, result);
-    }
-
-    void javaScriptConsoleMessage(const QString& message, int lineNumber, const QString& sourceID)
-    {
-        emit m_webPage->javaScriptConsoleMessageSent(message, lineNumber, sourceID);
-    }
-
-    void javaScriptError(const QString& message, int lineNumber, const QString& sourceID, const QString& stack)
-    {
-        emit m_webPage->javaScriptErrorSent(message, lineNumber, sourceID, stack);
-    }
-
-    void consoleMessageReceived(MessageSource source, MessageLevel level, const QString& message, int lineNumber, const QString& sourceID)
-    {
-        Q_UNUSED(source);
-
-        if (level == ErrorMessageLevel) {
-            emit m_webPage->javaScriptErrorSent(message, lineNumber, sourceID, QString());
-        } else {
-            emit m_webPage->javaScriptConsoleMessageSent(message, lineNumber, sourceID);
-        }
-    }
-
-    QString userAgentForUrl(const QUrl& url) const
-    {
-        Q_UNUSED(url);
-        return m_userAgent;
-    }
-
-    bool acceptNavigationRequest(QWebFrame* frame, const QNetworkRequest& request, QWebPage::NavigationType type)
-    {
-        bool isMainFrame = (frame == m_webPage->m_mainFrame);
-
-        QString navigationType = "Undefined";
-        switch (type) {
-        case NavigationTypeLinkClicked:
-            navigationType = "LinkClicked";
-            break;
-        case NavigationTypeFormSubmitted:
-            navigationType = "FormSubmitted";
-            break;
-        case NavigationTypeBackOrForward:
-            navigationType = "BackOrForward";
-            break;
-        case NavigationTypeReload:
-            navigationType = "Reload";
-            break;
-        case NavigationTypeFormResubmitted:
-            navigationType = "FormResubmitted";
-            break;
-        case NavigationTypeOther:
-            navigationType = "Other";
-            break;
-        }
-        bool isNavigationLocked = m_webPage->navigationLocked();
-
-        emit m_webPage->navigationRequested(
-            request.url().toEncoded(),       //< Requested URL
-            navigationType,                  //< Navigation Type
-            !isNavigationLocked,             //< Will navigate (not locked)?
-            isMainFrame);                    //< Is main frame?
-
-        return !isNavigationLocked;
-    }
-
-    QWebPage* createWindow(WebWindowType type)
-    {
-        Q_UNUSED(type);
-        WebPage* newPage;
-
-        // Create a new "raw" WebPage object
-        if (m_webPage->ownsPages()) {
-            newPage = new WebPage(m_webPage);
-        } else {
-            newPage = new WebPage(Phantom::instance());
-            Phantom::instance()->m_pages.append(newPage);
-        }
-        newPage->setCookieJar(m_cookieJar);
-
-        // Apply default settings
-        newPage->applySettings(Phantom::instance()->defaultPageSettings());
-
-        // Signal JS shim to catch, decorate and store this new child page
-        emit m_webPage->rawPageCreated(newPage);
-
-        // Return the new QWebPage to the QWebKit backend
-        return newPage->m_customWebPage;
-    }
-
-private:
-    WebPage* m_webPage;
-    QString m_userAgent;
-    QStringList m_uploadFiles;
-    friend class WebPage;
-    CookieJar* m_cookieJar;
-};
-
 
 /**
   * Contains the Callback Objects used to regulate callback-traffic from the webpage internal context.
@@ -270,13 +78,13 @@ class WebpageCallbacks : public QObject
     Q_OBJECT
 
 public:
-    WebpageCallbacks(QObject* parent = 0)
+    WebpageCallbacks(QObject* parent = Q_NULLPTR)
         : QObject(parent)
-        , m_genericCallback(NULL)
-        , m_filePickerCallback(NULL)
-        , m_jsConfirmCallback(NULL)
-        , m_jsPromptCallback(NULL)
-        , m_jsInterruptCallback(NULL)
+        , m_genericCallback(Q_NULLPTR)
+        , m_filePickerCallback(Q_NULLPTR)
+        , m_jsConfirmCallback(Q_NULLPTR)
+        , m_jsPromptCallback(Q_NULLPTR)
+        , m_jsInterruptCallback(Q_NULLPTR)
     {
     }
 
@@ -361,7 +169,7 @@ WebPage::WebPage(QObject* parent, const QUrl& baseUrl)
     setObjectName("WebPage");
     m_callbacks = new WebpageCallbacks(this);
     m_customWebPage = new CustomPage(this);
-    Config* phantomCfg = Phantom::instance()->config();
+    Settings* phantomCfg = Phantom::instance()->settings();
 
     // To grant universal access to a web page
     // attribute "WebSecurityEnabled" must be applied during the initializing
@@ -419,8 +227,7 @@ WebPage::WebPage(QObject* parent, const QUrl& baseUrl)
     if (isLocalStorageEnabled) {
         if (phantomCfg->localStoragePath().isEmpty()) {
             pageSettings->setLocalStoragePath(QStandardPaths::writableLocation(QStandardPaths::DataLocation));
-        }
-        else {
+        } else {
             pageSettings->setLocalStoragePath(phantomCfg->localStoragePath());
         }
     }
@@ -439,6 +246,9 @@ WebPage::WebPage(QObject* parent, const QUrl& baseUrl)
 
     m_dpi = qRound(QApplication::primaryScreen()->logicalDotsPerInch());
     m_customWebPage->setViewportSize(QSize(400, 300));
+
+    // create web page renderer
+    m_pageRenderer = new WebPageRenderer(m_customWebPage, m_dpi, this);
 }
 
 WebPage::~WebPage()
@@ -637,7 +447,7 @@ void WebPage::applySettings(const QVariantMap& def)
     opt->setAttribute(QWebSettings::JavascriptCanCloseWindows, def[PAGE_SETTINGS_JS_CAN_CLOSE_WINDOWS].toBool());
 
     if (def.contains(PAGE_SETTINGS_USER_AGENT)) {
-        m_customWebPage->m_userAgent = def[PAGE_SETTINGS_USER_AGENT].toString();
+        m_customWebPage->setUserAgent(def[PAGE_SETTINGS_USER_AGENT].toString());
     }
 
     if (def.contains(PAGE_SETTINGS_USERNAME)) {
@@ -679,7 +489,7 @@ void WebPage::setProxy(const QString& proxyUrl)
 
 QString WebPage::userAgent() const
 {
-    return m_customWebPage->m_userAgent;
+    return m_customWebPage->userAgent();
 }
 
 void WebPage::setNavigationLocked(bool lock)
@@ -761,7 +571,7 @@ QVariantMap WebPage::paperSize() const
 
 QVariant WebPage::evaluateJavaScript(const QString& code)
 {
-	QString function = "(" + code + ")()";
+    QString function = "(" + code + ")()";
 
     qDebug() << "WebPage - evaluateJavaScript" << function;
 
@@ -972,323 +782,53 @@ bool WebPage::render(const QString& fileName, const QVariantMap& option)
     if (m_mainFrame->contentsSize().isEmpty()) {
         return false;
     }
-
-    QString outFileName = fileName;
-    QString tempFileName = "";
-
     QString format = "";
-    int quality = -1; // QImage#save default
-
-    if (fileName == STDOUT_FILENAME || fileName == STDERR_FILENAME) {
-        if (!QFile::exists(fileName)) {
-            // create temporary file for OS that have no /dev/stdout or /dev/stderr. (ex. windows)
-            tempFileName = QDir::tempPath() + "/phantomjstemp" + QUuid::createUuid().toString();
-            outFileName = tempFileName;
-        }
-
-        format = "png"; // default format for stdout and stderr
-    } else {
-        QFileInfo fileInfo(outFileName);
-        QDir dir;
-        dir.mkpath(fileInfo.absolutePath());
-    }
+    int quality = -1;
 
     if (option.contains("format")) {
         format = option.value("format").toString();
     } else if (fileName.endsWith(".pdf", Qt::CaseInsensitive)) {
         format = "pdf";
+    } else {
+        // by default we use png
+        format = "png";
+    }
+    if (option.contains("quality")) {
+        quality = option.value("quality").toInt();
+    }
+
+    WebPageRenderer::RenderMode mode;
+    if (option.contains("onlyViewport") && option.value("onlyViewport").toBool()) {
+        mode = WebPageRenderer::Viewport;
+    } else {
+        mode = WebPageRenderer::Content;
+    }
+
+    return m_pageRenderer->renderToFile(fileName, format, m_clipRect, m_paperSize, mode, quality);
+}
+
+QString WebPage::renderBase64(const QByteArray& format, const QVariantMap& option)
+{
+    int quality = -1;
+
+    QByteArray outputFormat = format;
+
+    if (outputFormat.isEmpty()) {
+        outputFormat = "png";
     }
 
     if (option.contains("quality")) {
         quality = option.value("quality").toInt();
     }
 
-    bool retval = true;
-    if (format == "pdf") {
-        QPdfWriter pdfWriter(fileName);
-        retval = renderPdf(pdfWriter);
+    WebPageRenderer::RenderMode mode;
+    if (option.contains("onlyViewport") && option.value("onlyViewport").toBool()) {
+        mode = WebPageRenderer::Viewport;
     } else {
-        RenderMode mode;
-        if (option.contains("onlyViewport") && option.value("onlyViewport").toBool()) {
-            mode = Viewport;
-        } else {
-            mode = Content;
-        }
-        QImage rawPageRendering = renderImage(mode);
-
-        const char* f = 0; // 0 is QImage#save default
-        if (format != "") {
-            f = format.toLocal8Bit().constData();
-        }
-
-        retval = rawPageRendering.save(outFileName, f, quality);
+        mode = WebPageRenderer::Content;
     }
 
-    if (tempFileName != "") {
-        // cleanup temporary file and render to stdout or stderr
-        QFile i(tempFileName);
-        i.open(QIODevice::ReadOnly);
-
-        QByteArray ba = i.readAll();
-
-        System* system = (System*)Phantom::instance()->createSystem();
-        if (fileName == STDOUT_FILENAME) {
-#ifdef Q_OS_WIN
-            _setmode(_fileno(stdout), O_BINARY);
-#endif
-
-            ((File*)system->_stdout())->write(QString::fromLatin1(ba.constData(), ba.size()));
-
-#ifdef Q_OS_WIN
-            _setmode(_fileno(stdout), O_TEXT);
-#endif
-        } else if (fileName == STDERR_FILENAME) {
-#ifdef Q_OS_WIN
-            _setmode(_fileno(stderr), O_BINARY);
-#endif
-
-            ((File*)system->_stderr())->write(QString::fromLatin1(ba.constData(), ba.size()));
-
-#ifdef Q_OS_WIN
-            _setmode(_fileno(stderr), O_TEXT);
-#endif
-        }
-
-        i.close();
-
-        QFile::remove(tempFileName);
-    }
-
-    return retval;
-}
-
-QString WebPage::renderBase64(const QByteArray& format)
-{
-    QByteArray nformat = format.toLower();
-
-    if (nformat != "pdf" && !QImageWriter::supportedImageFormats().contains(nformat)) {
-        // Return an empty string in case an unsupported format was provided
-        return "";
-    }
-
-    // Prepare buffer for writing
-    QByteArray bytes;
-    QBuffer buffer(&bytes);
-    buffer.open(QIODevice::WriteOnly);
-
-    if (nformat == "pdf") {
-        QPdfWriter pdfWriter(&buffer);
-
-        if (!renderPdf(pdfWriter)) {
-            // Return an empty string if pdf render fails
-            return "";
-        }
-    } else {
-        QImage rawPageRendering = renderImage();
-
-        // Writing image to the buffer, using PNG encoding
-        rawPageRendering.save(&buffer, nformat);
-    }
-
-    return bytes.toBase64();
-}
-
-QImage WebPage::renderImage(const RenderMode mode)
-{
-    QRect frameRect;
-    QSize viewportSize = m_customWebPage->viewportSize();
-    if (mode == Viewport) {
-        frameRect = QRect(QPoint(0, 0), viewportSize);
-    } else {
-        QSize contentsSize = m_mainFrame->contentsSize();
-        contentsSize -= QSize(m_scrollPosition.x(), m_scrollPosition.y());
-        frameRect = QRect(QPoint(0, 0), contentsSize);
-        m_customWebPage->setViewportSize(contentsSize);
-    }
-
-    if (!m_clipRect.isNull()) {
-        frameRect = m_clipRect;
-    }
-
-#ifdef Q_OS_WIN
-    QImage::Format format = QImage::Format_ARGB32_Premultiplied;
-#else
-    QImage::Format format = QImage::Format_ARGB32;
-#endif
-
-    QImage buffer(frameRect.size(), format);
-    buffer.fill(Qt::transparent);
-
-    QPainter painter;
-
-    // We use tiling approach to work-around Qt software rasterizer bug
-    // when dealing with very large paint device.
-    // See http://code.google.com/p/phantomjs/issues/detail?id=54.
-    const int tileSize = 4096;
-    int htiles = (buffer.width() + tileSize - 1) / tileSize;
-    int vtiles = (buffer.height() + tileSize - 1) / tileSize;
-    for (int x = 0; x < htiles; ++x) {
-        for (int y = 0; y < vtiles; ++y) {
-
-            QImage tileBuffer(tileSize, tileSize, format);
-            tileBuffer.fill(Qt::transparent);
-
-            // Render the web page onto the small tile first
-            painter.begin(&tileBuffer);
-            painter.setRenderHint(QPainter::Antialiasing, true);
-            painter.setRenderHint(QPainter::TextAntialiasing, true);
-            painter.setRenderHint(QPainter::SmoothPixmapTransform, true);
-            painter.translate(-frameRect.left(), -frameRect.top());
-            painter.translate(-x * tileSize, -y * tileSize);
-            m_mainFrame->render(&painter, QRegion(frameRect));
-            painter.end();
-
-            // Copy the tile to the main buffer
-            painter.begin(&buffer);
-            painter.setCompositionMode(QPainter::CompositionMode_Source);
-            painter.drawImage(x * tileSize, y * tileSize, tileBuffer);
-            painter.end();
-        }
-    }
-    if (mode != Viewport) {
-        m_customWebPage->setViewportSize(viewportSize);
-    }
-    return buffer;
-}
-
-qreal WebPage::stringToPointSize(const QString& string) const
-{
-    static const struct {
-        QString unit;
-        qreal factor;
-    } units[] = {
-        { "mm", 72 / 25.4 },
-        { "cm", 72 / 2.54 },
-        { "in", 72 },
-        { "px", 72.0 / m_dpi },
-        { "", 72.0 / m_dpi }
-    };
-    for (uint i = 0; i < sizeof(units) / sizeof(units[0]); ++i) {
-        if (string.endsWith(units[i].unit)) {
-            QString value = string;
-            value.chop(units[i].unit.length());
-            return value.toDouble() * units[i].factor;
-        }
-    }
-    return 0;
-}
-
-qreal WebPage::printMargin(const QVariantMap& map, const QString& key)
-{
-    const QVariant margin = map.value(key);
-    if (margin.isValid() && margin.canConvert(QVariant::String)) {
-        return stringToPointSize(margin.toString());
-    } else {
-        return 0;
-    }
-}
-
-bool WebPage::renderPdf(QPdfWriter& pdfWriter)
-{
-    pdfWriter.setResolution(m_dpi);
-    QVariantMap paperSize = m_paperSize;
-
-    if (paperSize.isEmpty()) {
-        const QSize pageSize = m_mainFrame->contentsSize();
-        paperSize.insert("width", QString::number(pageSize.width()) + "px");
-        paperSize.insert("height", QString::number(pageSize.height()) + "px");
-        paperSize.insert("margin", "0px");
-    }
-
-    if (paperSize.contains("width") && paperSize.contains("height")) {
-        const QSizeF sizePt(ceil(stringToPointSize(paperSize.value("width").toString())),
-                            ceil(stringToPointSize(paperSize.value("height").toString())));
-        pdfWriter.setPageSize(QPageSize(sizePt, QPageSize::Point));
-    } else if (paperSize.contains("format")) {
-        const QPageLayout::Orientation orientation = paperSize.contains("orientation")
-                && paperSize.value("orientation").toString().compare("landscape", Qt::CaseInsensitive) == 0 ?
-                QPageLayout::Portrait : QPageLayout::Landscape;
-        pdfWriter.setPageOrientation(orientation);
-        static const struct {
-            QString format;
-            QPageSize::PageSizeId paperSize;
-        } formats[] = {
-            { "A0", QPageSize::A0 },
-            { "A1", QPageSize::A1 },
-            { "A2", QPageSize::A2 },
-            { "A3", QPageSize::A3 },
-            { "A4", QPageSize::A4 },
-            { "A5", QPageSize::A5 },
-            { "A6", QPageSize::A6 },
-            { "A7", QPageSize::A7 },
-            { "A8", QPageSize::A8 },
-            { "A9", QPageSize::A9 },
-            { "B0", QPageSize::B0 },
-            { "B1", QPageSize::B1 },
-            { "B2", QPageSize::B2 },
-            { "B3", QPageSize::B3 },
-            { "B4", QPageSize::B4 },
-            { "B5", QPageSize::B5 },
-            { "B6", QPageSize::B6 },
-            { "B7", QPageSize::B7 },
-            { "B8", QPageSize::B8 },
-            { "B9", QPageSize::B9 },
-            { "B10", QPageSize::B10 },
-            { "C5E", QPageSize::C5E },
-            { "Comm10E", QPageSize::Comm10E },
-            { "DLE", QPageSize::DLE },
-            { "Executive", QPageSize::Executive },
-            { "Folio", QPageSize::Folio },
-            { "Ledger", QPageSize::Ledger },
-            { "Legal", QPageSize::Legal },
-            { "Letter", QPageSize::Letter },
-            { "Tabloid", QPageSize::Tabloid }
-        };
-        pdfWriter.setPageSize(QPageSize(QPageSize::A4)); // Fallback
-        for (uint i = 0; i < sizeof(formats) / sizeof(formats[0]); ++i) {
-            if (paperSize.value("format").toString().compare(formats[i].format, Qt::CaseInsensitive) == 0) {
-                pdfWriter.setPageSize(QPageSize(formats[i].paperSize));
-                break;
-            }
-        }
-    } else {
-        return false;
-    }
-
-    if (paperSize.contains("border") && !paperSize.contains("margin")) {
-        // backwards compatibility
-        paperSize["margin"] = paperSize["border"];
-    }
-
-    qreal marginLeft = 0;
-    qreal marginTop = 0;
-    qreal marginRight = 0;
-    qreal marginBottom = 0;
-
-    if (paperSize.contains("margin")) {
-        const QVariant margins = paperSize["margin"];
-        if (margins.canConvert(QVariant::Map)) {
-            const QVariantMap map = margins.toMap();
-            marginLeft = printMargin(map, "left");
-            marginTop = printMargin(map, "top");
-            marginRight = printMargin(map, "right");
-            marginBottom = printMargin(map, "bottom");
-        } else if (margins.canConvert(QVariant::String)) {
-            const qreal margin = stringToPointSize(margins.toString());
-            marginLeft = margin;
-            marginTop = margin;
-            marginRight = margin;
-            marginBottom = margin;
-        }
-    }
-
-    pdfWriter.setPageMargins(QMarginsF(marginLeft, marginTop, marginRight, marginBottom), QPageLayout::Point);
-
-    QPainter painter(&pdfWriter);
-    m_mainFrame->render(&painter);
-    painter.end();
-
-    return true;
+    return m_pageRenderer->renderToBase64(outputFormat, m_clipRect, m_paperSize, mode, quality);
 }
 
 void WebPage::setZoomFactor(qreal zoom)
@@ -1316,7 +856,7 @@ qreal WebPage::getHeight(const QVariantMap& map, const QString& key) const
     if (!height.canConvert(QVariant::String)) {
         return 0;
     }
-    return stringToPointSize(height.toString());
+    return m_pageRenderer->stringToPointSize(height.toString());
 }
 
 qreal WebPage::footerHeight() const
@@ -1367,10 +907,10 @@ void WebPage::_uploadFile(const QString& selector, const QStringList& fileNames)
     }
 
     // Filter out "fileNames" that don't actually exist
-    m_customWebPage->m_uploadFiles.clear();
+    m_customWebPage->clearUploadFiles();
     for (int i = 0, ilen = fileNames.length(); i < ilen; ++i) {
         if (QFile::exists(fileNames[i])) {
-            m_customWebPage->m_uploadFiles.append(fileNames[i]);
+            m_customWebPage->addFileToUpload(fileNames[i]);
         }
     }
 
@@ -1379,7 +919,16 @@ void WebPage::_uploadFile(const QString& selector, const QStringList& fileNames)
 
 bool WebPage::injectJs(const QString& jsFilePath)
 {
-    return Utils::injectJsInFrame(jsFilePath, m_libraryPath, m_currentFrame);
+    QFile file(jsFilePath);
+
+    if (!file.open(QFile::ReadOnly | QFile::Text)) {
+        return false;
+    }
+
+    QTextStream in(&file);
+    m_currentFrame->evaluateJavaScript(in.readAll());
+
+    return true;
 }
 
 void WebPage::_appendScriptElement(const QString& scriptUrl)
